Covered PasswordService and PasswordOtpService with unit tests due to changes.
